day 5 part 2

+++ Day5/Day5.h
@@ -13,7 +13,7 @@ class Day5 {
 public:
     static int Part1();
 
-    static int Part2();
+    static long long Part2();
 
     static int Part3();
